fix client and some minor change
