+++ .gitignore
@@ -1,7 +1,4 @@
 __pycache__
-build/
 .vscode/
 deps/
-build_release/
-build_debug/
-bak/
+ppl-build/
